created cart context and updated the delete logic

+++ src/components/Cart/CartButton.tsx
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { ButtonGroup, Button } from 'react-bootstrap';
 import { MdAdd, MdRemove, MdDelete } from 'react-icons/md';
 import axios from 'axios';
+import { useCart } from '../../contexts/cartContext';
+import { useLocation } from 'react-router-dom';
 
 interface CartButtonProps {
   productId: number;
@@ -10,8 +12,12 @@ interface CartButtonProps {
 
 const CartButton: React.FC<CartButtonProps> = ({ productId, initialQuantity }) => {
   const [quantity, setQuantity] = useState<number>(initialQuantity);
+  const { handleCartUpdate } = useCart();
   const userInfo = localStorage.getItem('userInfo');
   const authToken = userInfo ? JSON.parse(userInfo).token : null;
+  const location = useLocation();
+
+  const isProductRoute = location.pathname.includes('/products');
 
   useEffect(() => {
     setQuantity(initialQuantity);
@@ -30,6 +36,7 @@ const CartButton: React.FC<CartButtonProps> = ({ productId, initialQuantity }) =
       );
       if (response.status === 200) {
         setQuantity((prevQuantity) => prevQuantity + 1);
+        handleCartUpdate(1); 
       } else {
         console.error('Failed to add item to cart');
       }
@@ -52,6 +59,7 @@ const CartButton: React.FC<CartButtonProps> = ({ productId, initialQuantity }) =
         );
         if (response.status === 200) {
           setQuantity((prevQuantity) => prevQuantity - 1);
+          handleCartUpdate(-1); 
         } else {
           console.error('Failed to update item quantity in cart');
         }
@@ -61,18 +69,25 @@ const CartButton: React.FC<CartButtonProps> = ({ productId, initialQuantity }) =
     }
   };
 
-  const handleDelete = async () => {
-  try {    
-    const response = await axios.delete(`http://localhost:4000/cart/delete/${productId}`, {
-      headers: {
-        Authorization: `Bearer ${authToken}`,
-      },
-    });
-    
-    if (response.status === 204) {
-      setQuantity(0);
+ const handleDelete = async () => {
+  try {
+    if (quantity === 1) {
+      // Perform delete action only when the quantity is 1
+      const response = await axios.delete(`http://localhost:4000/cart/delete/${productId}`, {
+        headers: {
+          Authorization: `Bearer ${authToken}`,
+        },
+      });
+      if (response.status === 204) {
+        setQuantity(0);
+        handleCartUpdate(-quantity); 
+      } else {
+        console.error('Failed to remove item from cart. Status:', response.status);
+      }
     } else {
-      console.error('Failed to remove item from cart. Status:', response.status);
+      // If quantity is greater than 1, perform item removal instead of delete
+      setQuantity((prevQuantity) => prevQuantity - 1);
+      handleCartUpdate(-1); 
     }
   } catch (error) {
     console.error('Error removing item from cart:', error);
@@ -80,33 +95,33 @@ const CartButton: React.FC<CartButtonProps> = ({ productId, initialQuantity }) =
 };
 
 
-
- return (
-  <ButtonGroup className="d-flex align-items-center w-100">
-    {quantity === 0 ? (
-      <Button variant="warning" onClick={handleAdd} className="align-items-center">
-        Add To Cart
-      </Button>
-    ) : (
-      <>
-        {quantity === 1 ? (
-          <Button variant="warning" onClick={handleDelete} className="align-items-center">
-            <MdDelete />
-          </Button>
-        ) : (
-          <Button variant="warning" onClick={handleRemove} className="align-items-center">
-            <MdRemove />
-          </Button>
-        )}
-        <Button variant="warning" disabled className="align-items-center">
-          {quantity}
-        </Button>
+  return (
+    <ButtonGroup className="d-flex align-items-center w-100">
+      {quantity === 0 ? (
         <Button variant="warning" onClick={handleAdd} className="align-items-center">
-          <MdAdd /> Add More
+          Add To Cart
         </Button>
-      </>
-    )}
-  </ButtonGroup>
-);
-}
+      ) : (
+        <>
+          {quantity === 1 ? (
+            <Button variant="warning" onClick={handleDelete} className="align-items-center">
+              <MdDelete />
+            </Button>
+          ) : (
+            <Button variant="warning" onClick={handleRemove} className="align-items-center">
+              <MdRemove />
+            </Button>
+          )}
+          <Button variant="warning" disabled className="align-items-center">
+            {quantity}
+          </Button>
+           <Button variant="warning" onClick={handleAdd} className="align-items-center">
+            <MdAdd /> {!isProductRoute && 'Add More'}
+          </Button>
+        </>
+      )}
+    </ButtonGroup>
+  );
+};
+
 export default CartButton;
